requirements.txt
openai>=1.51.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
numpy>=1.26.0
pandas>=2.2.3
duckduckgo-search>=7.4.0
ddgs>=9.0.0
requests>=2.32.3
beautifulsoup4>=4.12.3
trafilatura>=1.9.0
readability-lxml>=0.8.1
googlesearch-python>=1.2.5
tqdm>=4.67.1
datasets>=2.20.0
matplotlib>=3.10.0
sentence-transformers>=3.0.1
